Fixed an error with a project Boolean property always being turned back to true.
The ClearGdalCache property was not being written to project mwxml files if the value was false. However, absent ClearGdalCache properties are interpreted as being true. To address this, the ClearGdalCache property is always written to the project. The reason for this always being written is that ClearGdalCache is causing severe performance problems on raster layers when enabled. Keeping this potentially problematic property visible should assist developers and users in the future.